CLOUD-12325 - Bookmark share dialog
## Problem

We are adding bookmark sharing to Cloud which will allow a user to make a bookmark publicly visible to anyone (password optional).

This MR is adding the first phase of the share dialog which will appear after you click "Share" on a Bookmark

## Solution

Create a new dialog which was mostly an HTML and styling change. The Bookmark API has not been updated to support the V3/v4 versions with 6.0 features. That will need to be done before I can pull in any of the real share data here.

NOTE: This is all hidden behind a feature switch

| LIGHT | DARK |
| ------ | ------ |
|    ![image](/uploads/97a8ee4ee2e08fa8f27db45fc8ca7158/image.png)    |     ![image](/uploads/f5ab637733e963cb11964e7498b0fd63/image.png)   |