CLOUD-12571: Fix servers page not loading for offline system
# Problem
Servers page not loading for offline system.

# Solution
First issue was there was no error handling if the request to get servers failed.

Second issue was a redirect within the resolver, this seems to violate the single
responsibility rule and probably shouldn't be done but left it since I'm not sure
what cases it was meant to handle.

(cherry picked from commit 749f986c515410613907c6b2065bdc95dd18454d)