CLOUD-12439, CLOUD-12586: Filter out duplicate bookmarks
# Problem

`nxOnIntersect` triggers bookmarks fetch and sometimes it would fetch duplicate bookmarks. We need to filter out these duplicates

* Zooming in/out also triggers `nxOnIntersect`, which then triggers fetch
* We're getting duplicate bookmarks because of `fetchParams.endTimeMs = this.findOldestBookmark(this._bookmarks)?.startTimeMs - 1` - We use this for the fetch params so that we don't load new newer bookmarks as we scroll down the page. However, this will keep fetching bookmarks that start after `endTimeMs`, hence the duplicate bookmarks we're getting

# Changes

We'll have a list of unique bookmark IDs (`filteredFetchedBookmarkIds`). After we fetch bookmarks, we'll filter out the duplicate bookmarks by checking to see if the `bookmarkId` is already in `filteredFetchedBookmarkIds`

Related to CLOUD-12439