CLOUD-12648 - Rest V4 API and Bookmark Sharing additions
## Problem

For Bookmark Sharing we need to use the new RestV4 API for a few different endpoints. The first endpoint is the existing GET `/devices/*/bookmarks`, because it has an updated ID and a new field called "Share". In this MR we are creating the new V4 API service and using two of the endpoints to make sure everything is working properly

## Solution

* Create the new V4 API service and update the existing services to tell MediaServer to create one when it is 6.1
* Create the new get all bookmarks API in V4. Nothing changes here except adding v4 to the url
* Create a new Bookmark Type to account for the new Share field. **_ANY AND ALL SUGGESTIONS ON HOW TO IMPROVE THIS ARE WELCOME_**
* Update the bookmark sharing modal to PATCH in the share field when you click the share button