CLOUD-12547: Fixed routing to systems
# Problem
When you tried to navigate to a system the angular router was going to systems first and the route guard redirected it to home when the channel partners flag is turned on.

# Solution
Reorganize the routes so that all systems routes are together and move 'systems' below all other systems.

(cherry picked from commit 1957cb588b9527b6c5fc57b293817d750d39bdd6)