CLOUD-12762: Added exception handling for token validation
If an access token was expired when validation to check when it expired it thew an exception and resulted in the user getting logged out.

Added an exception handler for when token validation fails and set expires to 0 so that we try to refresh the access token.

(cherry picked from commit 022e56baa79b9a6167329f3b39ebfe2fcec9b436)