CLOUD-12671: Fix services placeholder page
## Problem
Dropdown did not handle values for `selected` that didn't match to any items, so it tried to scroll to a nonexistent item on open

## Solution
Treat no match like `undefined`. Tried intercepting and rejecting invalid values in `ngOnChanges`, but due to the way `QueryList` works the items aren't initialized at that time.