CLOUD-12330 - Bookmark Sharing - Password UI shows when password is required
## Problem

Bookmark Sharing will have an optional password. If there is no password then we will immediately show the Video Viewer page, but if a password is required then we should first show the password UI.

Here we are doing the logic to get the Bookmark Description, and if it errors we will change the page state to show the Password input components.

## Solution

We were already making a GET request to get the bookmark description. This will return with a few different errors if the password is not attached. Unfortunately it's impossible to know which error since they all use the same error codes. I have brought this up to the server team and hopefully they will address this. For now we will catch any error and if it is forbidden we show the password input box.

The design for these pages are heavily using the new Cloud 2.0 styling. Since we don't have any Cloud 2.0 components yet this is not heavily focused on the styling. Screenshots included to give you an idea of what the pages look like though.

| | **PASSWORD** | **VIEWER** |
| ------ | ------ | ------ |
| **LIGHT** | ![image](/uploads/76860b54a20a09f6d9f53824fc2fe66e/image.png) | ![image](/uploads/f5e22c85cb044dd368f9228bedc4810c/image.png) |
| **DARK** | ![image](/uploads/5eb20fd9d86d38b6943d573bce329222/image.png) | ![image](/uploads/d0fdf8bf7ba8414cf86b5e508522f825/image.png) |