CLOUD-11642: Update WebAdmin v1
A basic v1 for allowing users to update WebAdmin from the UI.

This MR creates a new Update WebAdmin page that is hidden (not in the menu, but can be found under `settings/update-webadmin`). It contains a form where you can enter a URL to download a new build, and an optional checksum. After submitting, the build will be downloaded in the backend, and the server's version of WebAdmin will update to that build when the download completes.

![image](/uploads/9b2082447875ecb40c61c310c4f63423/image.png){width=80%}

Next steps not included in this V1:

- "Reset to default build" button/section
- Don't allow the user to change to a version below the server default 
- Polling or JSON-RPC for download progress updates (right now just fetched once on page load or download initiation)
- Indicate required vs optional form fields in UI
- Improve error handling
  - When there is an issue with the URL in the backend
  - Form-level errors (system offline, etc)