CLOUD-12563 - Logic for text displaying when Bookmark Share expires
## Problem

For Bookmark Sharing we need to display the time remaining before a bookmark expires. The requirements from design are actually quite complicated, because the user will set the expiration time in generic intervals such as "Expires in a day" or "Expires in an hour". This complicated things, because when a user selects "Expires in an hour" it is immediately no longer an hour left (59 minutes, 59 seconds). So this handles that logic.

## Solution

The architecture for this solution is to create a function that will take in an expiration date and return an `Observable<Translatable>`. The reason it is an Observable is because we need the text to also be "live" and countdown (most notably when there are minutes/seconds remaining). To avoid unnecessary overhead we are only recalculating the time remaining when there is more than an hour left on the timer.

There is also a small refactor pulling this out into its own component. This will help make way for the edit dialog that is coming next


| Share Dialog | Edit Dialog | Edit Time Dropdown |
| ------ | ------ | ------ |
| ![image](/uploads/4e30be0c6041e8fabf16f8b1f17fff16/image.png) | ![image](/uploads/f08742f6be70d7d1cf52a846703e8d73/image.png) |    ![image](/uploads/6d0661e9fd7520b885d94d41a269ddcd/image.png) |