CLOUD-12652: Added network error handling to the account request.
# Problem
If the account poll encounters an error while fetching the user's account the app state breaks making it appear as though the user was logged out.

# Solution
We do what the app to break under certain conditions, but if the error status > 500 it's something with networking, and we should ignore it. If its any other status the lower catchError will handle the error effectively ending the poll.