CLOUD-11142, CLOUD-12630, CLOUD-12484: Ownership transfer updates
- Block transferring to org for <6.0 systems
- Refresh page on transfer to org (no accept action required)
- Update org transfer properties to signals and restructured template on admin
- Fixed system types

Figure out what to do about ownership information for org owned systems where the user is not in the org. They're in a weird position where since they're not user owned owner name doesn't come on the system response but since the user is not in the org they don't have permissions to get org name.