CLOUD-12330 - Bookmark Sharing password hash for password protected bookmarks
## Problem

For the Bookmark Sharing feature we have the ability to lock a bookmark behind a password. A user will share a bookmark and optionally include a password. We try to open the bookmark and if we get an error then we open the password prompt.

In this change we are getting the password input from the user and creating the password hash. We store this hash and send it with the request to get the bookmark data and the video data.

## Solution

When the user enters the password we do the formula provided by the API to generate the password hash. This formula requires the current server time (`synchronizedTimeMs + ":" + sha256(sha256(bookmarkId + password) + synchronizedTimeMs))`).

Current page flow:
1. When the page loads it tries to get the bookmark info
2. If the request returns unauthorized then there is a password and we simultaneously show the password input box and also make a request to get the server's current `synchronizedTimeMs` (`rest/v4/site/info`).
3. The user enters the password and clicks submit
4. We generate this password hash using the formula and store it.
5. We make the request again to get the bookmark info, but this time we pass the password hash.
6. If the request succeeds then we show the bookmark and make the request to get the video with the same password hash.

### Extras

* At the request of Nick I moved these components over to on push change detection to trial run what we should be doing for all components going forward if possible. This meant converting anything that changes the view over to a signal.
* Added the ability to submit the password by pressing enter