CLOUD-13024: Fixes for pre 6.0 systems
\# Problem\
The api call for creating a session cookie for rest systems had no error handling so the system guard would fail to finish running. When you click on the system tile a second time the system is offline so this step is skipped.\
\
# Solution\
Added a try catch for the session cookie api call. Now that the error is caught the system guard can finish and redirect the user to the system as expected.

(cherry picked from commit 56a6b9791edfa20e8a4b612991c9e6c57a16d3c3)