CLOUD-13068 - Dropdown Support for changing dropdown items
## Problem

The dropdown has an issue that when the currently selected option is changed, the placeholder does not get updated. An easy way to visualize this is with the sandbox demo I made, but that is obviously not a common use case (although in Bookmark Sharing it does happen).

The main time this will happen is if the currently selected option's text is added using the `async` pipe. So any observable or signal value that is set after the placeholder is set will not update the placeholder. This mostly happens when the option selected on component load uses an async value. In Bookmark Sharing this happens because our timer text is loaded async using the `getExpirationText` function

## Solution

The two solution options I could think of are:
* poll or somehow watch the children and refetch the innerHTML on any changes
* Make the innerHTML string that is passed instead be an observable and each dropdown item will push changes as needed

I went with the 2nd option. Each dropdown item will listen to character changes to its innerHTML using the `MutationObserver` and when it happens it will push a change to it's own `behaviorSubject`. Then instead of rendering a `string` of HTML we render an `observable<string>` of html

![Screen_Recording_2024-03-11_at_20.48.33](/uploads/d7184309cf80645619ac67bcda9bad89/Screen_Recording_2024-03-11_at_20.48.33.mov)