CLOUD-12332 - Bookmark Sharing - Add a 404 page when bookmark or server is not found
## Problem

Bookmark Sharing has a public page at the `/share/{systemId}/{bookmarkId}` url. Users could type in the wrong systemId or bookmarkId. In that case we should show a light weight 404 page.

## Solution

We don't know if the systemId or bookmarkId are incorrect until we make the request to get the bookmark info. At first I was going to catch all the error that could come back from this, but I think it's better to just show the 404 page anytime there is an error that is NOT the password error. Mostly we will get a 503 error if the systemId is wrong, or we will get an `internalServerError` if the bookmarkId is wrong.

Also fixed a small error where we weren't importing the translate module.

![image](/uploads/4f708276a7b97e6d9342eb83e1792cdc/image.png)