CLOUD-13164, CLOUD-13165, CLOUD-13170, CLOUD-13171: Batch of 6.0 user fixes
# CLOUD-13171: Fixes 6.0 user group names when offline

## Problem

CDB had changes that removed customPermissions for 6.0 system. Now it only provides roles and userRoleId.

## Solution

For 6.0 system add special logic to add default row info for offline systems. To make the systems resilient when offline the default groups were added to the codebase.

# CLOUD-13170: Fixes issue where Custom users didn't have accessRoles

## Problem

Roles can load later causing the user Access Role to be blank

## Solution

Return the users accessRole at the end of the function. If it's not truth return it as custom.

# 

# CLOUD-13164, CLOUD-13165: Fixed system initialization issues

## Problem

User groups do not load unless you are a power user, admin, or something with the current user changed in the system. Fixes for CLOUD-13024 9cc4a2c3bcee6e60957efc1ac575ff160c5d9224 caused parts of the system to not initialize when expected.

## Solution

Partially revert some of the changes and add it into the first poll for the system.

(cherry picked from commit 7d9c20be72abd51d79e2dc24c011211364e3b751)