CLOUD-13164, CLOUD-13165, CLOUD-13170, CLOUD-13171: Batch of 6.0 user fixes
CDB had changes that removed customPermissions for 6.0 system. Now it only provides roles and userRoleId.

For 6.0 system add special logic to add default row info for offline systems. To make the systems resilient when offline the default groups were added to the codebase.

Roles can load later causing the user Access Role to be blank

Return the users accessRole at the end of the function. If it's not truth return it as custom.

User groups do not load unless you are a power user, admin, or something with the current user changed in the system. Fixes for CLOUD-13024 9cc4a2c3bcee6e60957efc1ac575ff160c5d9224 caused parts of the system to not initialize when expected.

Partially revert some of the changes and add it into the first poll for the system.

(cherry picked from commit 7d9c20be72abd51d79e2dc24c011211364e3b751)