CLOUD-13093 - Bookmark Sharing - Share button changes to Shared when it's been shared
## Problem

In an effort to make the bookmark even more clean that it has been shared, design has updated the Share button to use a different icon and say "Shared" when a bookmark has already been share.

## Solution

* We can use the same solution we are using in a previous MR to determine if the bookmark has been shared. If you have question feel free to check the previous thread https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/6585#note_671942.
* Use a new checkmark icon if it has been shared and update the text

### Note:
I added the text to the static lang file. I'm not sure if this is necessary or if the `translate` pipe knows to grab this text to add it to the translation json. These 2 words are already in the compiled JSON, but I think it needs to be redundant in case whatever added `share` and `shared` get removed. We don't want to rely on the old components providing text to the new component


| Shared | Not Shared |
| ------ | ------ |
|    ![image](/uploads/16b4dcde104e93fb1c689333eb8fc127/image.png)    |   ![image](/uploads/96ef5d5c37cc821abc7f624dbc3a0418/image.png)     |