CLOUD-13107 - Bookmark Sharing - Show password error when incorrect password entered
## Problem

When entering a password for a password protected bookmark, there is currently no feedback to tell you that the password is incorrect. We need to add some error styling to show password was incorrect

## Solution

In a perfect world the Input here would be a shared component, but from what I can tell most of our inputs are not sharing a common base. Most of the work is done using global styles. There is a component called `NxAuthorizePasswordComponent` but it relies on the old `Process` pattern and is not worth using here. In the future we will hopefully make a Cloud 2.0 Input component, and also an extension for password inputs, that will do most of this logic/stying for us.

Anyway. In the meantime I am just copying from some other input fields and using their stylings for the error below the input. Bookmark Sharing will not release until 6.1 so we can replace this input field before it launches.

![image](/uploads/13bf554c5a6275ca7261686f764b3aa4/image.png)