CLOUD-12561  - Bookmark Sharing - adding a copied tooltip when user copies bookmark url
## Problem

For bookmark sharing you can copy a URL using a button, but there is no feedback to let the user know the copy was successful. We should add a tooltip that shows up when the user clicks.

## Solution

Utilize the existing tooltip directive. I needed to add one flag though so that the tooltip will go away after a certain duration. The idea is that it shows when the user clicks and then goes away on its own after 1.5s

![Screen_Recording_2024-03-22_at_13.03.35](/uploads/75b529fce7cf4ed5e1b6adbed91f6411/Screen_Recording_2024-03-22_at_13.03.35.mov)