CLOUD-12905, CLOUD-13052, CLOUD-13111, CLOUD-13000, CLOUD-13038, CLOUD-13190, CLOUD-12996: fix tab routing issues
# Problem

There are a lot of issues with routing within channel partners that along with some overly complicated logic within the tabs components was leading to the tabs being out of sync with the actual routes and also unpredictable behavior with navigating.

# Solution

Create a store that acts as the source of truth for the current router state within channel partners. Also provides a centralized place to handle generating routes. Currently all of the routing is done using click handlers instead of routerLinks and there's a bunch of places where the link is generated which leads to inconsistencies.