CLOUD-13309 - Scrollbar color issue
## Problem

The scrollbar color styling requires two colors to make it work. 1 for the scroll part and 1 for the scroll background. In its current form the scrollbar color just doesn't work at all.

This has existed for awhile but probably wasn't caught because of different browsers behaving differently

## Solution

add transparent to the scrollbar color property so that it gets picked up.

| BEFORE | AFTER |
| ------ | ------ |
|   ![image](/uploads/69a9dfa9992b737deee7d192b30f87e0/image.png)     |    ![image](/uploads/922216564b5a63f262b15e2088d7fc09/image.png)    |