CLOUD-13293, CLOUD-13297, and CLOUD-13327: Fix logic in updating org users
# Problem
Org users store wasn't updating, particularly when on parent /organizations route. It would request the users for the organization using the orgId from the router state store. At this point thought there isn't an active org, the request fails and since the exception wasn't caught it kills the observable.

# Solution
Add a switchMap to return NEVER when there isn't an active groupId/orgId. Also add a catchError and retry if the request for users fails for some reason.

Also fix redirect loop that was freezing cloud portal.