CLOUD-13235 - New Filters component (to replace existing Bookmark Filter)
## Problem

We have a few filters already for Bookmarks and they are not sharing any code. This was fine, but I needed to add another and it seemed like the right time to make this a real component. Especially because we have Advanced Object Search coming which will require a lot of filters, possibly even being dynamic.

#### Current filters

| Devices | Date/Time | Tags | Sharing | Advanced Object (future) | Advanced Object (future) |
| ------ | ------ | ------ | ------ | ------ | ------ |
| ![image](/uploads/2035da1be7f13e2a0816e383823ca7d4/image.png) | ![image](/uploads/41511c444e97454ad9e399f173586e4c/image.png) | ![image](/uploads/75b641c02c0bb4a9969441e92e0fce8d/image.png) | ![image](/uploads/3942a1cb39384c795f5fe2db70d14215/image.png) | ![image](/uploads/ebf28b22ea964662a484e647d5542c9e/image.png) | ![image](/uploads/00accfd0402a2c2a9fe8c6b92ca8381f/image.png) |

## Solution

As you can see they do share some commonalities in their formatting, and as we will need dynamic filtering we need to make something more robust.

Lets look first at how you use this, as I think this will help the most.

```html
<nx-filter-container header="Multiple States">
    <nx-single-item-filter [(ngModel)]="selectedState">
        @for (state of ALL_STATES; track state) {
            <nx-checkbox-filter-item
                [value]="state"
                [selected]="selectedState() === state"
            >
                {{ state }}
            </nx-checkbox-filter-item>
        }
    </nx-single-item-filter>
</nx-filter-container>
```

| `nx-filter-container` | `nx-single/multi-item-filter` | `nx-filter-item` |
| ------ | ------ | ------ |
| ![image](/uploads/edf5959e34a8877bc0760f8319513ced/image.png) | ![image](/uploads/70de3d514f62b2c2aede94d6dbf8d9a1/image.png) | ![image](/uploads/1779af3879e89cd41fa54558dc4f856a/image.png) |

I started with the base concept from the dropdownV2. You have a `nx-filter-container` which is just a structural and styling container. This also contains the little X button to clear the filter. In the future this may get more complex as we have an accordion style layout.

The main logic is houses in the `nx-single/multi-item-filter`. this is where the `ControlValueAccessor` for the filter is held and this is where you will get your currently selected filter items.

Lastly to build out the actual items in the filter we have `nx-filter-item`. These items are passed as the `ng-content` to a `nx-single/multi-item-filter`. They will hold the actual values that the filter items have and also be responsible for the styling. This isn't important now, but it will be in the future as you can see from the Advanced Object Search mockups.

### TODO Next
* Add the ability to have a limit on the amount shown (`displayLimit`)
* Add a button that says show more at the bottom of the `displayLimit` is reached. Still thinking on this. The filter can probably handle all of this logic, but I need to play with it to see if thats the best method.
* Support for the Tags style filter. "Chips" as they are sometimes called in the design world.
* Search bar. I have some ideas on this one that seem really fun and simple-ish
* Figure out the date filter. Maybe it is a filter that lives in this same Filter component directory, but probably will be mostly it's own logic. But can use the `nx-filter-container` at least and will refactor to use `ControlValueAccessor` like the others