CLOUD-13343: Fix playback issues
# Problem
Couldn't playback videos due to WebRTCStreamManager instances having issues accessing static properties on the class.
# Solution
Add a getter method on the instance to return reference to the class.

Need to look into what in our build is breaking the reference. It works fine within the example app and within 23.3.2
issue is only on develop and only on the production build.