CLOUD-13303: Updated the users component to get "live" updates
# Problem
Info for a user never updated because the value was set once in the router.

# Solution
Added a new observable for a user that is initialized from the router but receives live updates from the system with the poll.
There is a tradeoff to this solution. If a user is editing a setting for another user that setting will be overwritten by the update.

(cherry picked from commit 66c7f34cb06c72ebde1215a6448f1a99921a2a32)