CLOUD-12328 - Bookmark Sharing - Filter by Share or Not Shared
## Problem

We need to add a filter in bookmarks so that you can filter bookmarks by Shared and Not Shared. There was some previous work done to make a new filters component that will make this much easier.

## Solution

Utilize the new filter component to create a very simple filter for Share and Not Shared. Only one of these options can be selected at a time, hence we use the new radio buttons. The backend calls this `shareable` and `nonShareable` which are bad names, so there is a little bit of transforming that I did to use the words `shared` and `notShared`.

![trim-record-filter](/uploads/736bf972ec851fb30365b06d3e4c984f/trim-record-filter.mov)