CLOUD-13070: Fixed login dialog issues
# Problem
Due to how the dialog was being called it prevented users from logging into other rotes. What would happen is the local account service would create multiple instances of the dialog and redirect the user to /

# Solution
Added a subject to help debounce opening the login dialog. Next moved the result of the dialog into a observable so that the promise chain was less confusing to follow.