CLOUD-13346 CLOUD-13347 - Bookmark Sharing - Dropdown fix and Default Share fix
## Problem

Two bugs were raised for Bookmark Sharing
1. The default share time should be 1 day, but currently it is set to Never
2. When the expiration time is Never, it shows up twice in the dropdown

## Solution

1. This is super easy. We just change the expiration time from 0 to Date.now() + 1 day
2. The reason this happens is because we ALWAYS add an item to the dropdown for what the initial expiration value is. We can just not show that value if it is `0` aka never expires.


#### NOTE

There is still the issue where other options in the dropdown could be duplicated. This is because it's a more annoying issue to solve. We decided for now that we won't tackle that and will just solve this one case of Never Expires since its easy