CLOUD-13421: Signal Utilities and css color variables
# Changes
Add pipeSignal function to allow using rxjs operators to map signal state.

Add bindSignals to allow binding and mapping states of two signals.

Similar to bindSignals but binds to a single single and creates the second.

Add paramModel, creates a signal bound to a query param. Could also map to and from a param using a transformer.

Add paramSignal, similar to paramModel but readonly and only requires a deserializer instead of a transformer.

# Reason for the change
Simplify some complex patterns when working with signals.