CLOUD-13510: Add support for hiding licenses for org systems
# Summary
Added the ability to check if a system is in an organization and hide the licenses tab if it is in one. This is done by checking the `nxConfig.organizationId` property, which is set in the `NxBootstrapProvider`.

# Note
This needs to be handled differently on cloud, so it will be handled in CLOUD-12817.

(cherry picked from commit 3b1aa081becd45c05fd73c158004f019c56ebf3e)