CLOUD-13514 - Bookmark Sharing - Fix broken dropdown from previous change
## Problem

In a previous MR (https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/7027) we updated the dropdown name. Either this one was missed or it was pushed in afterwards.

## Solution

Change the HTML Tag so that it fixes the component. Currently the component doesn't load so just loading proves the fix worked.