CLOUD-13386 - Transfer Server to Org bug
There is a bug when a user tries to transfer a system to an Org, which is an Owner Action, and their session needs to be refreshed. The user will login, but the transfer request still fails.

This is because the transfer to an Org request requires the token to have a general scope, but our interceptor was refreshing the token to be scoped specifically to the system that is being transferred.

The interceptor will now accept a new token back from the Login Dialog instead of always grabbing it from the system.

The dialog will now either return a scoped token, a general token, or undefined. Previously the dialog just returned true or false if the login was successful.

This required a little bit of type changes as well.

(cherry picked from commit 7ff5fb1a025743348cedb62234d6a6f312da7cf1)