CLOUD-13583: Fix timeout on /rest/v1/login/sessions request
# Changes
Update PermissionManager.permissionsInitialized to fallback to getting the user to cloud if this request doesn't return after 3 seconds.

# Reason for the change
For some reason /rest/v1/login/sessions could potentially take a really long time to respond which can block the UI for a long time or if we use a timeout like we currently do it would load settings with the wrong permissions.