CLOUD-13516 CLOUD-13515 - Filters - Show More should only be visible when needed
## Problem

With the new Filter Component, we are always showing the "Show More" button and therefor always giving the user the popup. This is not needed for some filters, specifically filters that already have all of their options displayed in the sidebar.

## Solution

Add a `computed` that checks if the number of `FilterItems` is larger than the `displayLimit`. Then just add that as a check on when to show the button

| **BEFORE** | **AFTER** |
| ------ | ------ |
|    ![image](/uploads/efac15275e711022c7ebeb5c7f5985ee/image.png)    |    ![image](/uploads/f430f59283384f6b52be2e9118ab324a/image.png)    |