CLOUD-13449 CLOUD-13553 - New Organization Dialog - Validation changes
## Problem

Design has requested that the New Organization dialog have changes made to it's validation. They want the name field to be required, but for the user to not know that it is required until you click submit. Only then should be show an error.

## Solution

This requires a little bit of maneuvering because our global CSS and stuff gets applied only when the field is "touched". So what we do is:

* Make the required field false until the user clicks submit, then required turns into true.
* Check if the field is empty in the Process, if so then Promise.reject.
* When we catch the error. We have to make all the fields "touched" so that our global CSS will catch it.
* Also add an error message under the name so that it says "Organization Name is required". Design didnt ask for this, but it should be there to be consistent with all the other dialogs and their errors.

![image](/uploads/ac01505e486e96d98ef2bf8a0a2e629c/image.png)