CLOUD-13453 CLOUD-13576 CLOUD-13577 - Channel Partner Dialogs - New Folder/Rename Folder validation changes
## Problem

Design has requested that the New Folder and Rename Folder dialogs have changes made to it's validation. They want the name field to be required, but for the user to not know that it is required until you click submit. Only then should be show an error.

## Solution

This requires a little bit of maneuvering because our global CSS and stuff gets applied only when the field is "touched". So what we do is:

1. Validation changes
    - Make the required field false until the user clicks submit, then required turns into true.
    - Check if the field is empty in the Process, if so then Promise.reject.
    - When we catch the error. We have to make all the fields "touched" so that our global CSS will catch it.
2. Changed the label from `Name` -> `Folder Name`
3. Added the same 150 character limit to the Rename Folder dialog and added the error visualization to match


### Note

I moved both of these dialogs because they should have been in the channel partner folder.
Also, these probably should be combined into 1 dialog, but at this point it would be more work to do that. I think if we are gonna do that then we should look at refactoring a few of the dialogs into 1 component.

![image](/uploads/64cdca3f2b452bc51172eea2cb480410/image.png)