CLOUD-13452, CLOUD-13414: Update channel partners "not found" placeholders
We want to have one consistent placeholder when nothing is found when searching for organizations, users, sub-partners, or systems in all the different channel partners pages. This MR creates a "Nothing Found" version of the v2 page placeholder, and then uses this one component/version in all the places where we want to show the not found placeholder on search.

![image](/uploads/ebd4c45fb0ac4ff5b5ecd7ef6c69779c/image.png){width=70%}