CLOUD-13551, CLOUD-13552: Fixed issue with roles and offline systems
# Problem
The getUserRole method was broken with a previous change that was intended to fix the owner not showing up correctly when the system is offline. When it checked the owner role it could fail and go to other checks which effectively gives a user multiple chances to be the "owner"

# Solution
Refactor the getUserRole method so that we check userRoleId in the owner check to handle the case when the system is offline.

# Notes
Removed api calls for predefined roles, because they haven't changed since 4.2 and cdb's version of user info is compatible with pre 6.0 systems.