CLOUD-13606: Fixed issue where bind was broken if channel partners was off.
# Problem
The state for the bind step was initialized incorrectly so the preloader was showing, but users couldn't do anything.

# Solution
Update the default value for the bind step's state so that its undefined and then depending on if channel partners is enabled we start them in the initial state or confirmation state.

(cherry picked from commit b4a58be7f99ba5b7d4a34a0d8b40c07cc0311335)