CLOUD-13596: Fixed issue where users permissions didn't show correctly
# Problem
A bug caused permissions to fetch the users access role from the system. After the first time it's able to get the info from the system.

# Solution
Use resourceAccessRights from the system to determine if the user has custom or no permissions.

# Misc
Fixed issue where the web socket randomly dies. I opted to not use the retry operator because we'd need to change how the url is crafted.