CLOUD-10607: Change cookie policy to use agreements
# Changes
- Remove cookie checkbox from Account
- Create a new agreement type: cookie
- Allow the ContributorAgreemnt get_current function to filter for cookie type
- Change cookie banner to use cookie agreements

The new functionality is that if you are authenticated and the cookie policy is not accepted then you will be shown the banner. Once you click the X or accept the agreement on the agreement's page the banner will no longer show up. But, If you are unauthenticated and accept the cookie agreement, that does not actually do anything because an authenticated account must accept the agreement. The Cookie banner will still be hidden but will show back up if you authenticate with an account that has not accepted the cookie policy.