CLOUD-13560: Fixed redirect after logout
# Problem
When logging out of cloud the user is redirected to the landing page where it checks for the current user's name. Due to a race condition this field is still populated when we logged out from the header.

# Solution
When logging out of cloud we need to remove the current user from the store so that the landing page does redirect the user back to systems trigger the auth dialog to popup again.