CLOUD-12496 - New Bookmark search API for 6.0 and up
## Problem

Design wanted filtering by tags to be an inclusive search, so that we show all bookmarks that match ANY of the tags. This required the ability to do logic in our `text` search param. Backend has added this feature in 6.0 and up so we will alter how we send the request to accommodate that

## Solution

The backend now accepts the `AND` and `OR` keywords as well as parenthesis to separate the logic. Also, since all queries are case insensitive we should force the search queries to lower case that way it will never interfere with the logical keywords.

Examples:

* User searches for "Truck" and also selects the "Red" and "Blue" tags: `truck AND (red OR blue)`
* User does not search for anything but selects the "Red" tag: `red`
* User searches for "Truck" and selects the "Red" tag: `truck AND (red)`
    - this we could obviously do `truck AND red`, but it's no different and makes building the query less clunky