CLOUD-13596: Fixes issue where Custom permissions show as No permissions
# Problem
The permissionsInitialized method would get info from cdb if the current user wasn't populated when it was called. However, during that timeout a system can set the current user so the with method would override it and take info from cdb.

# Solution
Adjust the with pipe operator to check if the currentUser is ready one more time before resorting to cdb for the info.

(cherry picked from commit 3009509a8fc3c69cffec713f8f17b279f07124fa)