CLOUD-13610: Make Jenkins and Gitlab job statuses consistent
There is inconsistency between statuses of jobs in `Jenkins` and `Gitlab-ci`. A job that failed in `Jenkins` is displayed as successful in `Gitlab-ci`.
This is caused by putting a return code of job execution to a variable and not checking that variable further. So I've removed that from a pipe.

Additionally, using both stream redirect and `tee` command is excess. Those logs, created by tee are not stored anywhere so they're useless. Anyway all logs are visible in the console.