CLOUD-13518: Fix date issues in CP Reports
# Problem
There is a date bug in some CP report dates. When passing a date string (eg `2024-05-10`) to `dateFormat()`, the date is parsed as UTC midnight rather than the user's timezone, possibly causing the wrong day to show in the formatted result.

Described in more detail here
https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/7005#note_707410

# Solution
Parse the date string to construct the Date, to fix the timezone issue. Also use the new `NxDateTimeFormatService` instead of `formatDate` to work across locales.